[Refactoring] Move ConvertToTernaryExpr to its own file
